+++ core/app/app.go
@@ -21,7 +21,7 @@ var (
 	Name = "HORNET"
 
 	// Version of the app.
-	Version = "1.2.4"
+	Version = "1.2.5"
 )
 
 var (